Add exception handling
See [#2414347](https://www.drupal.org/node/2414347): when deleting items but Solr isn't running, it seems the Solr backend throws an exception from `deleteItems()` that is not a `SearchApiException` and, therefore, not permitted there. The exception should be caught and re-wrapped instead.

However, upon fixing this I saw that the code currently is never catching any Solarium exceptions (except in `ping()`), so this change is just a start, this needs to happen everywhere we use Solarium methods that can throw exceptions.